project/app/trial/requirements.txt
Flask==2.0.1
requests==2.26.0
gunicorn==20.1.0
jaeger-client==4.8.0
opentelemetry-api==1.5.0
opentelemetry-sdk==1.5.0
opentelemetry-exporter-jaeger==1.5.0
opentelemetry-instrumentation==0.24b0
opentelemetry-instrumentation-flask==0.24b0
opentelemetry-instrumentation-requests==0.24b0
